Correcion DB.sql y cambiar las entidades double a decimal

+++ lib_dominio/Entidades/Productos.cs
@@ -10,10 +10,10 @@ namespace lib_dominio.Entidades
     {
         public int ID { get; set; }
         public string? Nombre { get; set; }
-        public double Precio_Compra { get; set; }
+        public decimal Precio_Compra { get; set; }
         public int? Cant_Embase { get; set; }
         public string? Descripcion { get; set; }
-        public double Precio_Venta { get; set; }
+        public decimal Precio_Venta { get; set; }
         public int? Stock { get; set; }
 
         //public List<Productos_Documentos>? ProductosDocumentos { get; set; }
